setup.sh defaults to emacs, but replaces with specific emacs-NAME if it exists
git-svn-id: http://svn.dgtized.net/main/home-config@76 fb327865-0606-0410-b31b-95a48d5bc805

+++ setup.sh
@@ -8,7 +8,12 @@ function main () {
     ln -sfv ${DOTC_DIR}/site-config ~/.site-config
     
     ln -sfv ${DOTC_DIR}/bashrc ~/.bashrc
-    ln -sfv ${DOTC_DIR}/emacs-${DOTC_NAME} ~/.emacs
+    if [[ -e ${DOTC_DIR}/emacs-${DOTC_NAME} ]]; then
+    	ln -sfv ${DOTC_DIR}/emacs-${DOTC_NAME} ~/.emacs
+    else
+	ln -sfv ${DOTC_DIR}/emacs ~/.emacs
+    fi
+
     ln -sfv ${DOTC_DIR}/vimrc ~/.vimrc
     
     ln -sfv ${DOTC_DIR}/irbrc ~/.irbrc
